+++ Sources/KeyboardKit/Styling/KeyboardStyle+Background.swift
@@ -10,34 +10,28 @@ import SwiftUI
 
 public extension KeyboardStyle {
     
-    /**
-     This general background style can be used to define the
-     background of a keyboard or any components within it.
-     
-     The style has many layers, which should be applied from
-     the bottom to the top, starting with a background color
-     and gradient, then an image, then finally some overlays.
-     All these layers are optional.
-     
-     You can use the ``backgroundView`` property to create a
-     view with all the layers, applied in the intended order.
-     
-     The ``standard`` style value can be used to get and set
-     the global default style.
-     */
-    struct Background: Codable, Equatable {
+    /// This background style can define the background of a
+    /// keyboard or any components within.
+    ///
+    /// This style has many optional layers, which should be
+    /// added from the bottom to the top. A background color,
+    /// a gradient, an image, then some overlays.
+    ///
+    /// Since the style implements `View`, you can render it
+    /// directly into any view hierarchy.
+    ///
+    /// You can use the ``standard`` style or your own style.
+    struct Background: Codable, Equatable, View {
         
-        /**
-         Create a background style with optional components.
-         
-         - Parameters:
-           - backgroundColor: A background color to apply, by default `nil`.
-           - backgroundGradient: A background gradient color set, by default `nil`.
-           - imageData: Optional image data for a background image, by default `nil`.
-           - imageContentMode: The background image content mode, by default `.fill`.
-           - overlayColor: An overlay color to apply, by default `nil`.
-           - overlayGradient: An overlay gradient color set, by default `nil`.
-         */
+        /// Create a keyboard button background style.
+        ///
+        /// - Parameters:
+        ///   - backgroundColor: A background color to apply, by default `nil`.
+        ///   - backgroundGradient: A background gradient color set, by default `nil`.
+        ///   - imageData: Optional image data for a background image, by default `nil`.
+        ///   - imageContentMode: The background image content mode, by default `.fill`.
+        ///   - overlayColor: An overlay color to apply, by default `nil`.
+        ///   - overlayGradient: An overlay gradient color set, by default `nil`.
         public init(
             backgroundColor: Color? = nil,
             backgroundGradient: [Color]? = nil,
@@ -61,9 +55,9 @@ public extension KeyboardStyle {
             
             var nativeMode: ContentMode? {
                 switch self {
-                case .stretch: return nil
-                case .fill: return .fill
-                case .fit: return .fit
+                case .stretch: nil
+                case .fill: .fill
+                case .fit: .fit
                 }
             }
         }
@@ -85,25 +79,45 @@ public extension KeyboardStyle {
         
         /// An overlay gradient color set.
         public var overlayGradient: [Color]?
+        
+        
+        public var body: some View {
+            ZStack {
+                backgroundColor
+                if let backgroundGradient {
+                    LinearGradient(colors: backgroundGradient, startPoint: .top, endPoint: .bottom)
+                }
+                image(from: imageData)?
+                    .resizable()
+                    .optionalAspectRatio(with: imageContentMode)
+                if let overlayGradient {
+                    LinearGradient(colors: overlayGradient, startPoint: .top, endPoint: .bottom)
+                }
+                overlayColor
+            }
+        }
     }
 }
 
 public extension KeyboardStyle.Background {
 
-    /**
-     The standard button border style.
-
-     This can be changed to affect the global, default style.
-     */
+    /// The standard keyboard background style.
+    ///
+    /// You can set this style to affect the global default.
     static var standard = Self()
 
-    /// Create a background style with a single color.
-    static func color(_ color: Color) -> Self {
+    /// A background style with a single color.
+    static func color(
+        _ color: Color
+    ) -> Self {
         .init(backgroundColor: color)
     }
     
-    /// Create a background style with a single image.
-    static func image(_ data: Data, contentMode: ImageContentMode = .fill) -> Self {
+    /// A background style with a single image.
+    static func image(
+        data: Data,
+        contentMode: ImageContentMode = .fill
+    ) -> Self {
         .init(
             imageData: data,
             imageContentMode: contentMode
@@ -111,44 +125,39 @@ public extension KeyboardStyle.Background {
     }
     
     #if os(iOS) || os(tvOS) || os(watchOS)
-    /// Create a background style with a single image.
-    static func image(_ image: UIImage?, contentMode: ImageContentMode = .fill) -> Self {
+    /// A background style with a single image.
+    static func image(
+        _ image: UIImage?,
+        contentMode: ImageContentMode = .fill
+    ) -> Self {
         .init(
             imageData: image?.pngData() ?? Data(),
             imageContentMode: contentMode
         )
     }
     
-    /// Create a background style with a single image.
-    static func image(systemName: String, contentMode: ImageContentMode = .fill) -> Self {
+    /// A background style with a single image.
+    static func image(
+        systemName: String,
+        contentMode: ImageContentMode = .fill
+    ) -> Self {
         .image(
-            UIImage(systemName: "face.smiling"),
+            UIImage(systemName: systemName),
             contentMode: contentMode
         )
     }
     #endif
 
-    /// Create a background style with a vertical gradient.
-    static func verticalGradient(_ colors: [Color]) -> Self {
+    /// A background style with a vertical gradient.
+    static func verticalGradient(
+        _ colors: [Color]
+    ) -> Self {
         .init(backgroundGradient: colors)
     }
 
-    /// Create a background view with all style properties.
-    @ViewBuilder
+    @available(*, deprecated, renamed: "body", message: "This style now implements View and can be used right away")
     var backgroundView: some View {
-        ZStack {
-            backgroundColor
-            if let backgroundGradient {
-                LinearGradient(colors: backgroundGradient, startPoint: .top, endPoint: .bottom)
-            }
-            image(from: imageData)?
-                .resizable()
-                .optionalAspectRatio(with: imageContentMode)
-            if let overlayGradient {
-                LinearGradient(colors: overlayGradient, startPoint: .top, endPoint: .bottom)
-            }
-            overlayColor
-        }
+        body
     }
 }
 
@@ -182,17 +191,12 @@ private extension KeyboardStyle.Background {
     }
 }
 
-struct KeyboardStyle_Background_Previews: PreviewProvider {
+#Preview {
     
-    static var previews: some View {
-        VStack {
-            KeyboardButton.Key(style: .preview1)
-            KeyboardButton.Key(style: .preview2)
-            KeyboardButton.Key(style: .previewImage)
-        }
-        .padding()
-        .background(Color.gray)
-        .cornerRadius(10)
-        .environment(\.sizeCategory, .extraExtraLarge)
+    VStack {
+        KeyboardStyle.Background.color(.red)
+        KeyboardStyle.Background.verticalGradient([.red, .blue])
+        KeyboardStyle.Background.image(systemName: "checkmark")
     }
+    .padding()
 }